Updated implementation.

+++ Analytics.Model/SalesForecasting/IProductSales.cs
@@ -4,6 +4,6 @@ namespace Analytics.Model.SalesForecasting
 {
     public interface IProductSales
     {
-        Task<ProductUnitPrediction> Predict(string modelPath, string productId, int year, int month, float units, float avg, int count, float max, float min, float prev);
+        Task<ProductUnitPrediction> Predict(string modelPath, string productId, float year, float month, float units, float avg, int count, float max, float min, float prev);
     }
 }
